refactor(PathFinder): extract class
- Path 생성 책임이 과도하게 커짐에 따라 별도의 클래스로 분리
